Fix warning with some versions of gcc.
	* src/console-server.cc: static cast when doing longlong =
	  long+longlong.

git-svn-id: file:///srv/svn/kernel1/trunk@2131 1e37f2df-8004-0410-ba96-adcf012cb36e

+++ src/console-server.cc
@@ -50,7 +50,7 @@ public:
   }
   virtual void beforeWork()
   {
-    ctime += period_get() * 1000LL;
+    ctime += static_cast<long long>(period_get()) * 1000LL;
   }
   virtual void reset()
   {}
